pyproject.toml
[project]
name = "api-gateway-auth"
version = "0.1.0"
description = ""
authors = [
    {name = "Wellyson Balbino",email = "wellysonjose115@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi[standard] (>=0.133.0,<0.134.0)",
    "redis (>=7.2.0,<8.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "pyjwt (>=2.11.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "api_gateway_auth", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
